- Move `NodeCacheTestUtils` to `msal-node-samples` folder.
