+++ samples/msal-node-samples/client-credentials/test/client-credentials-aad.spec.ts
@@ -1,5 +1,5 @@
 import { LabApiQueryParams } from "../../../e2eTestUtils/LabApiQueryParams";
-import { NodeCacheTestUtils } from "../../../e2eTestUtils/NodeCacheTestUtils";
+import { NodeCacheTestUtils } from "../../NodeCacheTestUtils";
 import {RETRY_TIMES, retrieveAppConfiguration} from "../../../e2eTestUtils/TestUtils";
 import { LabClient } from "../../../e2eTestUtils/LabClient";
 import { validateCacheLocation } from "../../testUtils";
