+++ samples/msal-node-samples/package.json
@@ -16,6 +16,7 @@
     },
     "dependencies": {
         "@types/express": "^4.17.11",
+        "@azure/msal-node": "file:../../lib/msal-node",
         "express": "^4.17.1"
     }
 }
