+++ samples/msal-node-samples/silent-flow/test/silent-flow-b2c-local.spec.ts
@@ -12,7 +12,7 @@ import {
     ONE_SECOND_IN_MS,
     RETRY_TIMES
 } from "../../../e2eTestUtils/TestUtils";
-import { NodeCacheTestUtils } from "../../../e2eTestUtils/NodeCacheTestUtils";
+import { NodeCacheTestUtils } from "../../NodeCacheTestUtils";
 import { LabClient } from "../../../e2eTestUtils/LabClient";
 import { LabApiQueryParams } from "../../../e2eTestUtils/LabApiQueryParams";
 import { B2cProviders, UserTypes } from "../../../e2eTestUtils/Constants";
